pyproject.toml
[tool.poetry]
name = "medicalchatbot"
version = "0.1.0"
description = ""
authors = ["NIKHIL KUMAR REDDY <nikhilkumar.m@skil.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
instructor = "^1.3.7"
fastapi = {extras = ["all"], version = "^0.112.1"}
pydantic-settings = "^2.4.0"
bhashini-translator = "^0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
